# docs/spec/openapi.yml
openapi: 3.0.2
servers:
  - url: /v1
info:
  version: 1.0.0
  title: Percona Everest schema
...
paths:
  '/kubernetes':
    post:
      tags:
        - k8s
      summary: Register kubernetes cluster in Everest
      description: Register kubernetes cluster in Everest
      operationId: registerKubernetesCluster
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KubernetesCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The kubernetes cluster object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKubernetesClusterParams'
    get:
      tags:
        - k8s
      summary: List of the registered kubernetes clusters
      description: List of the registered kubernetes clusters
      operationId: listKubernetesClusters
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KubernetesClusterList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/kubernetes/{kubernetes-id}':
    get:
      tags:
        - k8s
      summary: Get the specified kubernetes cluster
      description: Get the specified kubernetes cluster
      operationId: getKubernetesCluster
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KubernetesCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/kubernetes/{kubernetes-id}/database-clusters':
    post:
      tags:
        - databaseCluster
      summary: Create a database cluster on the specified kubernetes cluster
      description: Create a database cluster on the specified kubernetes cluster
      operationId: createDatabaseCluster
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCluster'
    get:
      tags:
        - databaseCluster
      summary: List of the created database clusters on the specified kubernetes cluster
      description: List of the created database clusters on the specified kubernetes cluster
      operationId: listDatabaseClusters
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/kubernetes/{kubernetes-id}/database-clusters/{name}':
    get:
      tags:
        - databaseCluster
      summary: Get the specified database cluster on the specified kubernetes cluster
      description: Get the specified database cluster on the specified kubernetes cluster
      operationId: getDatabaseCluster
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - databaseCluster
      summary: Replace the specified database cluster on the specified kubernetes cluster
      description: Replace the specified database cluster on the specified kubernetes cluster
      operationId: updateDatabaseCluster
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCluster'
    delete:
      tags:
        - databaseCluster
      summary: Delete the specified database cluster on the specified kubernetes cluster
      description: Delete the specified database cluster on the specified kubernetes cluster
      operationId: deleteDatabaseCluster
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/kubernetes/{kubernetes-id}/database-engines':
    get:
      tags:
        - databaseEngine
      summary: List of the available database engines on the specified kubernetes cluster
      description: List of available database engines on the specified kubernetes cluster
      operationId: listDatabaseEngines
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseEngineList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/kubernetes/{kubernetes-id}/database-engines/{name}':
    get:
      tags:
        - databaseEngine
      summary: Get the specified database engine on the specified kubernetes cluster
      description: Get the specified database engine on the specified kubernetes cluster
      operationId: getDatabaseEngine
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database engine
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseEngineList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/kubernetes/{kubernetes-id}/database-cluster-restores':
    post:
      tags:
        - databaseClusterRestore
      summary: Create a database cluster restore on the specified kubernetes cluster
      description: Create a database cluster restore on the specified kubernetes cluster
      operationId: createDatabaseClusterRestore
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '201':
          description: Created success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster restore object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseClusterRestore'
    get:
      tags:
        - databaseClusterRestore
      summary: List of the created database cluster restores on the specified kubernetes cluster
      description: List of the created database cluster restores on the specified kubernetes cluster
      operationId: listDatabaseClusterRestores
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestoreList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/kubernetes/{kubernetes-id}/database-cluster-restores/{name}':
    get:
      tags:
        - databaseClusterRestore
      summary: Returns the specified cluster restore on the specified kubernetes cluster
      description: Returns the specified cluster restore on the specified kubernetes cluster
      operationId: getDatabaseClusterRestore
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster restore. Can be found under Metadata["name"] of the DatabaseClusterRestore object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - databaseClusterRestore
      summary: Replace the specified cluster restore on the specified kubernetes cluster
      description: Replace the specified cluster restore on the specified kubernetes cluster
      operationId: updateDatabaseClusterRestore
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster restore. Can be found under Metadata["name"] of the DatabaseClusterRestore object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '201':
          description: Created successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster restore object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseClusterRestore'
    delete:
      tags:
        - databaseClusterRestore
      summary: Delete the specified cluster restore on the specified kubernetes cluster
      description: Delete the specified cluster restore on the specified kubernetes cluster
      operationId: deleteDatabaseClusterRestore
      parameters:
        - name: kubernetes-id
          in: path
          description: Id of the kubernetes cluster
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster restore. Can be found under Metadata["name"] of the DatabaseClusterRestore object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
    CreateKubernetesClusterParams:
      type: object
      description: kubernetes object
      properties:
        name:
          type: string
          example: minikube
        kubeconfig:
          type: string
      required:
        - name
        - kubeconfig
    KubernetesCluster:
      type: object
      description: kubernetes object
      properties:
        id:
          type: string
        name:
          type: string
          example: minikube
    KubernetesClusterList:
      type: array
      items:
        type: object
        $ref: '#/components/schemas/KubernetesCluster'
    DatabaseClusterList:
      description: DatabaseClusterList is an object that contains the list of the existing database clusters.
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation
                  of an object. Servers should convert recognized schemas to the latest
                  internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        kind:
          description: 'Kind is a string value representing the REST resource this
                  object represents. Servers may infer this from the endpoint the client
                  submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        items:
          type: array
          $ref: '#/components/schemas/DatabaseCluster'
        metadata:
          type: object
      type: object
    DatabaseEngineList:
      description: DatabaseEngineList is an object that contains the list of the existing database engines.
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation
                  of an object. Servers should convert recognized schemas to the latest
                  internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        kind:
          description: 'Kind is a string value representing the REST resource this
                  object represents. Servers may infer this from the endpoint the client
                  submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        items:
          type: array
          $ref: '#/components/schemas/DatabaseEngine'
        metadata:
          type: object
      type: object
    DatabaseClusterRestoreList:
      description: DatabaseClusterRestoreList is an object that contains the list of the existing database cluster restores.
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation
                  of an object. Servers should convert recognized schemas to the latest
                  internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        kind:
          description: 'Kind is a string value representing the REST resource this
                  object represents. Servers may infer this from the endpoint the client
                  submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        items:
          type: array
          $ref: '#/components/schemas/DatabaseClusterRestore'
        metadata:
          type: object
      type: object
...
    io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2:
      description: Status is a return value for calls that don't return other objects.
      type: object
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation
            of an object. Servers should convert recognized schemas to the latest internal
            value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        code:
          description: Suggested HTTP return code for this status, 0 if not set.
          type: integer
          format: int32
        details:
          description: Extended data associated with the reason.  Each reason may define
            its own extended details. This field is optional and the data returned is
            not guaranteed to conform to any schema except that defined by the reason
            type.
          "$ref": "#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.StatusDetails_v2"
        kind:
          description: 'Kind is a string value representing the REST resource this object
            represents. Servers may infer this from the endpoint the client submits requests
            to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        message:
          description: A human-readable description of the status of this operation.
          type: string
        metadata:
          description: 'Standard list metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          "$ref": "#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.ListMeta"
        reason:
          description: A machine-readable description of why this operation is in the
            "Failure" status. If this value is empty there is no information available.
            A Reason clarifies an HTTP status code but does not override it.
          type: string
        status:
          description: 'Status of the operation. One of: "Success" or "Failure". More
            info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status'
          type: string

    # -------------------------
    # io.k8s.apimachinery.pkg.apis.meta.v1.ListMeta spec is imported from k8s API,
    # please do not edit the definition below manually
    # -------------------------
...
    io.k8s.apimachinery.pkg.apis.meta.v1.StatusDetails_v2:
      description: StatusDetails is a set of additional properties that MAY be set by
        the server to provide additional information about a response. The Reason field
        of a Status object defines what attributes will be set. Clients must ignore fields
        that do not match the defined type of each attribute, and should assume that any
        attribute may be empty, invalid, or under defined.
      type: object
      properties:
        causes:
          description: The Causes array includes more details associated with the StatusReason
            failure. Not all StatusReasons may provide detailed causes.
          type: array
          items:
            "$ref": "#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.StatusCause"
        group:
          description: The group attribute of the resource associated with the status
            StatusReason.
          type: string
        kind:
          description: 'The kind attribute of the resource associated with the status
            StatusReason. On some operations may differ from the requested resource Kind.
            More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        name:
          description: The name attribute of the resource associated with the status StatusReason
            (when there is a single name which can be described).
          type: string
        retryAfterSeconds:
          description: If specified, the time in seconds before the operation should be
            retried. Some errors may indicate the client must take an alternate action
            - for those errors this field may indicate how long to wait before taking
            the alternate action.
          type: integer
          format: int32
        uid:
          description: 'UID of the resource. (when there is a single resource which can
            be described). More info: http://kubernetes.io/docs/user-guide/identifiers#uids'
          type: string

    # -------------------------
    # io.k8s.apimachinery.pkg.apis.meta.v1.StatusCause spec is imported from k8s API,
    # please do not edit the definition below manually
    # -------------------------
...